Move _InputOutput in socket package
